feat(refactor): Refactor db server from vanilla js to typescript and add folder for db files
