+++ .gitignore
@@ -1,3 +1,5 @@
 .idea
 node_modules
 .DS_Store
+dist
+db
